+++ app/src/main/java/com/isodev/cocktailappui/data/model/CocktailCategoryModel.kt
@@ -1,0 +1,7 @@
+package com.isodev.cocktailappui.data.model
+
+data class CocktailCategoryModel(
+    val cocktailName: String,
+    val cocktailImage: Int,
+    val cocktailMixes: String,
+)
