+++ app/src/main/java/com/isodev/cocktailappui/ui/navigation/Pager.kt
@@ -1,0 +1,47 @@
+package com.isodev.cocktailappui.ui.navigation
+
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.pager.HorizontalPager
+import androidx.compose.foundation.pager.rememberPagerState
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.util.lerp
+import com.isodev.cocktailappui.ui.component.LayeredCardDesign
+import kotlin.math.absoluteValue
+
+@OptIn(ExperimentalFoundationApi::class)
+@Composable
+fun Pager(modifier: Modifier = Modifier) {
+    val pagerState = rememberPagerState(pageCount = { 4 })
+
+    HorizontalPager(
+        state = pagerState,
+        contentPadding = PaddingValues(horizontal = 75.dp),
+    ) { page ->
+        val pageOffset = ((pagerState.currentPage - page) + pagerState.currentPageOffsetFraction).absoluteValue
+
+        val offsetY = lerp(
+            start = 80f,
+            stop = 0f,
+            fraction = pageOffset.coerceIn(0f, 1f)
+        )
+
+        val scale = lerp(
+            start = 0.75f,
+            stop = 1f,
+            fraction = 1f - pageOffset.coerceIn(0f, 1f)
+        )
+
+        LayeredCardDesign(
+            modifier = modifier
+                .graphicsLayer {
+                    scaleX = scale
+                    scaleY = scale
+                },
+            offsetY = offsetY
+        )
+    }
+}
