--- app/src/main/res/drawable/i_want_learn_text.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="191dp"
-    android:height="19dp"
-    android:viewportWidth="191"
-    android:viewportHeight="19">
-  <path
-      android:pathData="M4.02,1.25V18H0.66V1.25H4.02ZM31.49,4.7L27.6,18H23.97L21.55,8.71L19.13,18H15.48L11.57,4.7H14.97L17.33,14.83L19.87,4.7H23.42L25.92,14.81L28.27,4.7H31.49ZM32.44,11.3C32.44,9.96 32.7,8.77 33.23,7.73C33.78,6.69 34.5,5.89 35.42,5.33C36.34,4.77 37.38,4.49 38.51,4.49C39.5,4.49 40.37,4.69 41.1,5.09C41.86,5.49 42.46,5.99 42.9,6.6V4.7H46.29V18H42.9V16.06C42.47,16.68 41.87,17.2 41.1,17.62C40.35,18.02 39.48,18.22 38.49,18.22C37.37,18.22 36.34,17.93 35.42,17.35C34.5,16.78 33.78,15.97 33.23,14.93C32.7,13.87 32.44,12.66 32.44,11.3ZM42.9,11.35C42.9,10.54 42.74,9.84 42.42,9.26C42.1,8.67 41.67,8.22 41.13,7.92C40.58,7.6 40,7.44 39.38,7.44C38.75,7.44 38.18,7.59 37.65,7.9C37.12,8.2 36.69,8.65 36.35,9.24C36.03,9.82 35.87,10.5 35.87,11.3C35.87,12.1 36.03,12.81 36.35,13.42C36.69,14.01 37.12,14.46 37.65,14.78C38.19,15.1 38.77,15.26 39.38,15.26C40,15.26 40.58,15.11 41.13,14.81C41.67,14.49 42.1,14.04 42.42,13.46C42.74,12.87 42.9,12.17 42.9,11.35ZM56.94,4.51C58.52,4.51 59.8,5.02 60.78,6.02C61.75,7.02 62.24,8.41 62.24,10.2V18H58.88V10.66C58.88,9.6 58.62,8.79 58.09,8.23C57.56,7.66 56.84,7.37 55.93,7.37C55,7.37 54.27,7.66 53.72,8.23C53.19,8.79 52.93,9.6 52.93,10.66V18H49.57V4.7H52.93V6.36C53.38,5.78 53.95,5.34 54.63,5.02C55.34,4.68 56.11,4.51 56.94,4.51ZM69.35,7.46V13.9C69.35,14.34 69.45,14.67 69.66,14.88C69.89,15.07 70.25,15.17 70.77,15.17H72.33V18H70.21C67.38,18 65.97,16.62 65.97,13.87V7.46H64.38V4.7H65.97V1.42H69.35V4.7H72.33V7.46H69.35ZM84.37,7.46V13.9C84.37,14.34 84.48,14.67 84.68,14.88C84.91,15.07 85.28,15.17 85.79,15.17H87.35V18H85.24C82.4,18 80.99,16.62 80.99,13.87V7.46H79.4V4.7H80.99V1.42H84.37V4.7H87.35V7.46H84.37ZM95.67,18.22C94.39,18.22 93.24,17.94 92.21,17.38C91.19,16.8 90.38,15.99 89.79,14.95C89.21,13.91 88.93,12.71 88.93,11.35C88.93,9.99 89.22,8.79 89.81,7.75C90.42,6.71 91.25,5.91 92.29,5.35C93.33,4.78 94.49,4.49 95.77,4.49C97.05,4.49 98.21,4.78 99.25,5.35C100.29,5.91 101.1,6.71 101.69,7.75C102.3,8.79 102.61,9.99 102.61,11.35C102.61,12.71 102.29,13.91 101.67,14.95C101.06,15.99 100.23,16.8 99.17,17.38C98.13,17.94 96.97,18.22 95.67,18.22ZM95.67,15.29C96.28,15.29 96.85,15.14 97.37,14.86C97.92,14.55 98.35,14.1 98.67,13.51C98.99,12.92 99.15,12.2 99.15,11.35C99.15,10.09 98.81,9.12 98.14,8.45C97.49,7.76 96.68,7.42 95.72,7.42C94.76,7.42 93.95,7.76 93.29,8.45C92.65,9.12 92.33,10.09 92.33,11.35C92.33,12.62 92.65,13.59 93.27,14.28C93.91,14.95 94.71,15.29 95.67,15.29ZM114.15,0.24V18H110.79V0.24H114.15ZM129.83,11.06C129.83,11.54 129.8,11.98 129.73,12.36H120.01C120.09,13.32 120.43,14.07 121.02,14.62C121.61,15.16 122.34,15.43 123.21,15.43C124.45,15.43 125.34,14.9 125.87,13.82H129.49C129.11,15.1 128.37,16.16 127.29,16.99C126.2,17.81 124.86,18.22 123.28,18.22C122,18.22 120.85,17.94 119.82,17.38C118.81,16.8 118.02,15.99 117.44,14.95C116.89,13.91 116.6,12.71 116.6,11.35C116.6,9.98 116.89,8.77 117.44,7.73C118,6.69 118.79,5.89 119.8,5.33C120.81,4.77 121.96,4.49 123.28,4.49C124.54,4.49 125.67,4.76 126.66,5.3C127.67,5.85 128.45,6.62 128.99,7.63C129.55,8.62 129.83,9.77 129.83,11.06ZM126.35,10.1C126.33,9.24 126.02,8.55 125.41,8.04C124.81,7.51 124.06,7.25 123.18,7.25C122.35,7.25 121.64,7.5 121.07,8.02C120.51,8.51 120.17,9.21 120.04,10.1H126.35ZM131.42,11.3C131.42,9.96 131.68,8.77 132.21,7.73C132.75,6.69 133.48,5.89 134.39,5.33C135.32,4.77 136.35,4.49 137.49,4.49C138.48,4.49 139.35,4.69 140.08,5.09C140.83,5.49 141.43,5.99 141.88,6.6V4.7H145.26V18H141.88V16.06C141.45,16.68 140.85,17.2 140.08,17.62C139.33,18.02 138.46,18.22 137.46,18.22C136.35,18.22 135.32,17.93 134.39,17.35C133.48,16.78 132.75,15.97 132.21,14.93C131.68,13.87 131.42,12.66 131.42,11.3ZM141.88,11.35C141.88,10.54 141.72,9.84 141.4,9.26C141.08,8.67 140.65,8.22 140.1,7.92C139.56,7.6 138.98,7.44 138.35,7.44C137.73,7.44 137.15,7.59 136.63,7.9C136.1,8.2 135.66,8.65 135.33,9.24C135.01,9.82 134.85,10.5 134.85,11.3C134.85,12.1 135.01,12.81 135.33,13.42C135.66,14.01 136.1,14.46 136.63,14.78C137.17,15.1 137.74,15.26 138.35,15.26C138.98,15.26 139.56,15.11 140.1,14.81C140.65,14.49 141.08,14.04 141.4,13.46C141.72,12.87 141.88,12.17 141.88,11.35ZM151.91,6.77C152.34,6.06 152.9,5.51 153.59,5.11C154.29,4.71 155.09,4.51 155.99,4.51V8.04H155.1C154.04,8.04 153.24,8.29 152.7,8.78C152.17,9.28 151.91,10.14 151.91,11.38V18H148.55V4.7H151.91V6.77ZM165.62,4.51C167.2,4.51 168.48,5.02 169.46,6.02C170.43,7.02 170.92,8.41 170.92,10.2V18H167.56V10.66C167.56,9.6 167.3,8.79 166.77,8.23C166.24,7.66 165.52,7.37 164.61,7.37C163.68,7.37 162.95,7.66 162.4,8.23C161.87,8.79 161.61,9.6 161.61,10.66V18H158.25V4.7H161.61V6.36C162.06,5.78 162.63,5.34 163.31,5.02C164.02,4.68 164.79,4.51 165.62,4.51ZM175.6,18.17C175,18.17 174.49,17.98 174.09,17.62C173.71,17.23 173.52,16.76 173.52,16.2C173.52,15.64 173.71,15.18 174.09,14.81C174.49,14.42 175,14.23 175.6,14.23C176.2,14.23 176.68,14.42 177.07,14.81C177.45,15.18 177.65,15.64 177.65,16.2C177.65,16.76 177.45,17.23 177.07,17.62C176.68,17.98 176.2,18.17 175.6,18.17ZM181.84,18.17C181.23,18.17 180.73,17.98 180.33,17.62C179.94,17.23 179.75,16.76 179.75,16.2C179.75,15.64 179.94,15.18 180.33,14.81C180.73,14.42 181.23,14.23 181.84,14.23C182.43,14.23 182.92,14.42 183.3,14.81C183.69,15.18 183.88,15.64 183.88,16.2C183.88,16.76 183.69,17.23 183.3,17.62C182.92,17.98 182.43,18.17 181.84,18.17ZM188.07,18.17C187.47,18.17 186.96,17.98 186.56,17.62C186.18,17.23 185.99,16.76 185.99,16.2C185.99,15.64 186.18,15.18 186.56,14.81C186.96,14.42 187.47,14.23 188.07,14.23C188.67,14.23 189.15,14.42 189.54,14.81C189.92,15.18 190.11,15.64 190.11,16.2C190.11,16.76 189.92,17.23 189.54,17.62C189.15,17.98 188.67,18.17 188.07,18.17Z"
-      android:fillColor="#1E2742"/>
-</vector>
