+++ app/src/main/res/drawable/icon_cocktail.xml
