+++ app/src/main/res/drawable/icon_cocktail_star.xml
