--- app/src/main/res/drawable/recent_mixes.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="112dp"
-    android:height="15dp"
-    android:viewportWidth="112"
-    android:viewportHeight="15">
-  <path
-      android:pathData="M0.44,14V1.22H6C6.58,1.22 7.11,1.34 7.59,1.58C8.07,1.82 8.48,2.14 8.83,2.53C9.19,2.93 9.46,3.37 9.66,3.87C9.86,4.36 9.96,4.86 9.96,5.38C9.96,5.94 9.86,6.48 9.66,7C9.45,7.51 9.16,7.96 8.77,8.33C8.4,8.7 7.96,8.97 7.46,9.14L10.47,14H8.2L5.43,9.55H2.46V14H0.44ZM2.46,7.77H5.98C6.36,7.77 6.69,7.66 6.97,7.45C7.26,7.23 7.49,6.94 7.66,6.58C7.83,6.21 7.91,5.81 7.91,5.38C7.91,4.93 7.81,4.54 7.62,4.19C7.43,3.83 7.18,3.54 6.87,3.33C6.57,3.11 6.24,3 5.88,3H2.46V7.77ZM16.06,14.18C15.32,14.18 14.66,14.05 14.06,13.8C13.47,13.54 12.96,13.18 12.53,12.74C12.1,12.28 11.76,11.77 11.52,11.19C11.29,10.6 11.18,9.98 11.18,9.32C11.18,8.43 11.38,7.62 11.79,6.89C12.2,6.15 12.77,5.55 13.5,5.11C14.23,4.65 15.09,4.42 16.07,4.42C17.06,4.42 17.9,4.65 18.61,5.11C19.33,5.55 19.89,6.14 20.29,6.87C20.68,7.6 20.88,8.39 20.88,9.23C20.88,9.37 20.87,9.51 20.86,9.64C20.85,9.76 20.84,9.87 20.83,9.95H13.28C13.32,10.5 13.47,10.99 13.73,11.41C14.01,11.82 14.36,12.14 14.78,12.38C15.2,12.61 15.65,12.72 16.13,12.72C16.66,12.72 17.15,12.59 17.62,12.33C18.1,12.06 18.43,11.71 18.59,11.28L20.29,11.77C20.08,12.22 19.77,12.64 19.35,13.01C18.94,13.37 18.46,13.66 17.89,13.87C17.33,14.08 16.72,14.18 16.06,14.18ZM13.23,8.64H18.92C18.88,8.1 18.73,7.62 18.45,7.21C18.19,6.79 17.84,6.47 17.42,6.24C17.02,6 16.56,5.88 16.06,5.88C15.56,5.88 15.11,6 14.69,6.24C14.28,6.47 13.94,6.79 13.68,7.21C13.42,7.62 13.27,8.1 13.23,8.64ZM26.76,14.18C26.03,14.18 25.36,14.05 24.76,13.8C24.17,13.54 23.66,13.18 23.23,12.72C22.8,12.27 22.46,11.74 22.22,11.16C22,10.57 21.88,9.94 21.88,9.28C21.88,8.4 22.08,7.59 22.48,6.85C22.88,6.12 23.45,5.53 24.17,5.09C24.9,4.65 25.76,4.42 26.76,4.42C27.72,4.42 28.55,4.64 29.26,5.07C29.98,5.49 30.51,6.06 30.85,6.76L28.92,7.38C28.7,6.98 28.4,6.67 28,6.46C27.62,6.23 27.19,6.12 26.72,6.12C26.2,6.12 25.72,6.25 25.28,6.53C24.86,6.79 24.53,7.17 24.28,7.65C24.04,8.11 23.92,8.66 23.92,9.28C23.92,9.88 24.04,10.43 24.29,10.92C24.55,11.4 24.88,11.79 25.3,12.07C25.73,12.35 26.21,12.49 26.74,12.49C27.07,12.49 27.38,12.43 27.68,12.33C27.99,12.21 28.26,12.05 28.49,11.86C28.72,11.65 28.88,11.43 28.97,11.19L30.92,11.77C30.71,12.24 30.41,12.65 30,13.01C29.6,13.37 29.13,13.66 28.58,13.87C28.04,14.08 27.43,14.18 26.76,14.18ZM36.73,14.18C36,14.18 35.33,14.05 34.73,13.8C34.14,13.54 33.63,13.18 33.2,12.74C32.77,12.28 32.43,11.77 32.19,11.19C31.96,10.6 31.85,9.98 31.85,9.32C31.85,8.43 32.05,7.62 32.46,6.89C32.87,6.15 33.44,5.55 34.17,5.11C34.9,4.65 35.76,4.42 36.75,4.42C37.73,4.42 38.58,4.65 39.28,5.11C40,5.55 40.56,6.14 40.96,6.87C41.35,7.6 41.55,8.39 41.55,9.23C41.55,9.37 41.55,9.51 41.53,9.64C41.52,9.76 41.51,9.87 41.5,9.95H33.96C33.99,10.5 34.14,10.99 34.41,11.41C34.68,11.82 35.03,12.14 35.45,12.38C35.87,12.61 36.32,12.72 36.8,12.72C37.33,12.72 37.83,12.59 38.29,12.33C38.77,12.06 39.1,11.71 39.27,11.28L40.96,11.77C40.75,12.22 40.44,12.64 40.02,13.01C39.61,13.37 39.13,13.66 38.56,13.87C38,14.08 37.39,14.18 36.73,14.18ZM33.9,8.64H39.59C39.55,8.1 39.4,7.62 39.12,7.21C38.86,6.79 38.52,6.47 38.1,6.24C37.69,6 37.23,5.88 36.73,5.88C36.24,5.88 35.78,6 35.36,6.24C34.95,6.47 34.62,6.79 34.35,7.21C34.09,7.62 33.94,8.1 33.9,8.64ZM51.61,14H49.63V8.73C49.63,7.85 49.48,7.21 49.2,6.8C48.92,6.39 48.51,6.19 47.95,6.19C47.57,6.19 47.19,6.28 46.8,6.48C46.43,6.67 46.09,6.93 45.79,7.27C45.49,7.59 45.28,7.97 45.15,8.4V14H43.17V4.59H44.97V6.48C45.21,6.06 45.52,5.7 45.9,5.4C46.3,5.08 46.74,4.84 47.23,4.68C47.73,4.51 48.25,4.42 48.8,4.42C49.36,4.42 49.83,4.53 50.19,4.73C50.56,4.93 50.85,5.22 51.05,5.58C51.27,5.92 51.41,6.33 51.48,6.8C51.57,7.26 51.61,7.74 51.61,8.26V14ZM59.07,13.51C58.91,13.6 58.7,13.69 58.44,13.78C58.19,13.88 57.9,13.96 57.58,14.04C57.27,14.11 56.94,14.14 56.61,14.14C56.2,14.14 55.82,14.07 55.47,13.93C55.13,13.78 54.84,13.56 54.63,13.24C54.42,12.92 54.32,12.51 54.32,12V6.12H53.08V4.59H54.32V1.49H56.3V4.59H58.35V6.12H56.3V11.37C56.33,11.7 56.43,11.94 56.61,12.09C56.8,12.25 57.03,12.33 57.31,12.33C57.59,12.33 57.85,12.28 58.1,12.18C58.35,12.09 58.54,12.01 58.66,11.95L59.07,13.51ZM75.89,14V4.89L72.13,11.8H70.94L67.16,4.89V14H65.15V1.22H67.31L71.54,9.03L75.77,1.22H77.93V14H75.89ZM80.56,14V4.59H82.54V14H80.56ZM80.56,3.04V0.86H82.54V3.04H80.56ZM86.03,4.59L88.26,7.81L88.42,8.11L88.59,7.81L90.8,4.59H92.94L89.56,9.3L92.98,14H90.84L88.59,10.8L88.42,10.51L88.26,10.8L86.01,14H83.87L87.29,9.32L83.91,4.59H86.03ZM98.15,14.18C97.41,14.18 96.75,14.05 96.15,13.8C95.56,13.54 95.05,13.18 94.62,12.74C94.19,12.28 93.85,11.77 93.61,11.19C93.38,10.6 93.27,9.98 93.27,9.32C93.27,8.43 93.47,7.62 93.88,6.89C94.29,6.15 94.86,5.55 95.59,5.11C96.32,4.65 97.18,4.42 98.16,4.42C99.15,4.42 99.99,4.65 100.7,5.11C101.42,5.55 101.98,6.14 102.38,6.87C102.77,7.6 102.97,8.39 102.97,9.23C102.97,9.37 102.96,9.51 102.95,9.64C102.94,9.76 102.93,9.87 102.92,9.95H95.37C95.41,10.5 95.56,10.99 95.82,11.41C96.1,11.82 96.45,12.14 96.87,12.38C97.29,12.61 97.74,12.72 98.22,12.72C98.75,12.72 99.24,12.59 99.71,12.33C100.19,12.06 100.51,11.71 100.68,11.28L102.38,11.77C102.17,12.22 101.86,12.64 101.44,13.01C101.03,13.37 100.54,13.66 99.98,13.87C99.42,14.08 98.81,14.18 98.15,14.18ZM95.32,8.64H101.01C100.97,8.1 100.82,7.62 100.54,7.21C100.28,6.79 99.93,6.47 99.51,6.24C99.11,6 98.65,5.88 98.15,5.88C97.65,5.88 97.2,6 96.78,6.24C96.37,6.47 96.03,6.79 95.77,7.21C95.51,7.62 95.36,8.1 95.32,8.64ZM107.93,14.18C107.14,14.18 106.39,14.05 105.68,13.8C104.97,13.54 104.37,13.15 103.86,12.63L104.6,11.3C105.14,11.78 105.68,12.13 106.22,12.34C106.77,12.56 107.32,12.67 107.86,12.67C108.41,12.67 108.86,12.57 109.21,12.36C109.56,12.15 109.73,11.83 109.73,11.43C109.73,11.13 109.64,10.9 109.46,10.76C109.28,10.62 109.02,10.49 108.67,10.38C108.33,10.27 107.93,10.15 107.45,10.02C106.77,9.82 106.21,9.61 105.75,9.39C105.3,9.16 104.96,8.89 104.73,8.56C104.5,8.24 104.39,7.83 104.39,7.32C104.39,6.71 104.54,6.19 104.85,5.76C105.17,5.32 105.6,4.99 106.15,4.77C106.71,4.54 107.36,4.42 108.09,4.42C108.78,4.42 109.41,4.53 109.98,4.75C110.57,4.95 111.09,5.28 111.53,5.72L110.69,7.03C110.27,6.65 109.83,6.37 109.39,6.19C108.96,6.01 108.49,5.92 107.99,5.92C107.7,5.92 107.42,5.95 107.16,6.03C106.89,6.09 106.67,6.21 106.49,6.39C106.32,6.55 106.24,6.79 106.24,7.11C106.24,7.36 106.31,7.56 106.46,7.72C106.6,7.87 106.82,8.01 107.1,8.11C107.39,8.22 107.75,8.34 108.17,8.47C108.9,8.69 109.52,8.91 110.04,9.12C110.55,9.33 110.95,9.6 111.23,9.93C111.51,10.26 111.66,10.71 111.66,11.28C111.66,12.19 111.32,12.91 110.63,13.42C109.96,13.93 109.06,14.18 107.93,14.18Z"
-      android:fillColor="#1E2742"/>
-</vector>
