--- app/src/main/res/drawable/see_all.xml
@@ -1,15 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="82dp"
-    android:height="54dp"
-    android:viewportWidth="82"
-    android:viewportHeight="54">
-  <path
-      android:pathData="M25,20.484C24.92,20.396 24.804,20.304 24.652,20.208C24.5,20.104 24.32,20.008 24.112,19.92C23.904,19.832 23.68,19.76 23.44,19.704C23.2,19.648 22.952,19.62 22.696,19.62C22.136,19.62 21.72,19.724 21.448,19.932C21.176,20.14 21.04,20.428 21.04,20.796C21.04,21.068 21.116,21.284 21.268,21.444C21.42,21.596 21.652,21.724 21.964,21.828C22.276,21.932 22.668,22.044 23.14,22.164C23.716,22.3 24.216,22.464 24.64,22.656C25.064,22.84 25.388,23.088 25.612,23.4C25.836,23.712 25.948,24.124 25.948,24.636C25.948,25.068 25.868,25.44 25.708,25.752C25.548,26.064 25.324,26.32 25.036,26.52C24.748,26.712 24.416,26.856 24.04,26.952C23.664,27.04 23.26,27.084 22.828,27.084C22.404,27.084 21.98,27.04 21.556,26.952C21.14,26.864 20.74,26.736 20.356,26.568C19.98,26.392 19.632,26.18 19.312,25.932L19.924,24.78C20.028,24.884 20.176,25 20.368,25.128C20.568,25.248 20.796,25.364 21.052,25.476C21.316,25.588 21.6,25.684 21.904,25.764C22.216,25.836 22.532,25.872 22.852,25.872C23.388,25.872 23.796,25.78 24.076,25.596C24.364,25.404 24.508,25.132 24.508,24.78C24.508,24.492 24.416,24.264 24.232,24.096C24.048,23.92 23.78,23.772 23.428,23.652C23.076,23.532 22.656,23.412 22.168,23.292C21.608,23.14 21.14,22.976 20.764,22.8C20.388,22.616 20.104,22.384 19.912,22.104C19.728,21.816 19.636,21.452 19.636,21.012C19.636,20.452 19.772,19.98 20.044,19.596C20.316,19.204 20.684,18.912 21.148,18.72C21.62,18.52 22.148,18.42 22.732,18.42C23.116,18.42 23.48,18.464 23.824,18.552C24.168,18.632 24.488,18.744 24.784,18.888C25.088,19.032 25.364,19.196 25.612,19.38L25,20.484ZM29.972,27.12C29.484,27.12 29.04,27.036 28.64,26.868C28.248,26.692 27.908,26.456 27.62,26.16C27.332,25.856 27.108,25.512 26.948,25.128C26.796,24.736 26.72,24.32 26.72,23.88C26.72,23.288 26.856,22.748 27.128,22.26C27.4,21.764 27.78,21.368 28.268,21.072C28.756,20.768 29.328,20.616 29.984,20.616C30.64,20.616 31.204,20.768 31.676,21.072C32.156,21.368 32.528,21.76 32.792,22.248C33.056,22.736 33.188,23.26 33.188,23.82C33.188,23.916 33.184,24.008 33.176,24.096C33.168,24.176 33.16,24.244 33.152,24.3H28.124C28.148,24.668 28.248,24.992 28.424,25.272C28.608,25.544 28.84,25.76 29.12,25.92C29.4,26.072 29.7,26.148 30.02,26.148C30.372,26.148 30.704,26.06 31.016,25.884C31.336,25.708 31.552,25.476 31.664,25.188L32.792,25.512C32.656,25.816 32.448,26.092 32.168,26.34C31.896,26.58 31.572,26.772 31.196,26.916C30.82,27.052 30.412,27.12 29.972,27.12ZM28.088,23.424H31.88C31.856,23.064 31.752,22.748 31.568,22.476C31.392,22.196 31.164,21.98 30.884,21.828C30.612,21.668 30.308,21.588 29.972,21.588C29.644,21.588 29.34,21.668 29.06,21.828C28.788,21.98 28.564,22.196 28.388,22.476C28.212,22.748 28.112,23.064 28.088,23.424ZM37.109,27.12C36.621,27.12 36.177,27.036 35.777,26.868C35.385,26.692 35.045,26.456 34.757,26.16C34.469,25.856 34.245,25.512 34.085,25.128C33.933,24.736 33.857,24.32 33.857,23.88C33.857,23.288 33.993,22.748 34.265,22.26C34.537,21.764 34.917,21.368 35.405,21.072C35.893,20.768 36.465,20.616 37.121,20.616C37.777,20.616 38.341,20.768 38.813,21.072C39.293,21.368 39.665,21.76 39.929,22.248C40.193,22.736 40.325,23.26 40.325,23.82C40.325,23.916 40.321,24.008 40.313,24.096C40.305,24.176 40.297,24.244 40.289,24.3H35.261C35.285,24.668 35.385,24.992 35.561,25.272C35.745,25.544 35.977,25.76 36.257,25.92C36.537,26.072 36.837,26.148 37.157,26.148C37.509,26.148 37.841,26.06 38.153,25.884C38.473,25.708 38.689,25.476 38.801,25.188L39.929,25.512C39.793,25.816 39.585,26.092 39.305,26.34C39.033,26.58 38.709,26.772 38.333,26.916C37.957,27.052 37.549,27.12 37.109,27.12ZM35.225,23.424H39.017C38.993,23.064 38.889,22.748 38.705,22.476C38.529,22.196 38.301,21.98 38.021,21.828C37.749,21.668 37.445,21.588 37.109,21.588C36.781,21.588 36.477,21.668 36.197,21.828C35.925,21.98 35.701,22.196 35.525,22.476C35.349,22.748 35.249,23.064 35.225,23.424ZM46.619,18.48H47.891L51.203,27H49.787L48.887,24.684H45.599L44.711,27H43.283L46.619,18.48ZM48.623,23.688L47.255,19.956L45.839,23.688H48.623ZM52.089,18.24H53.409V25.188C53.409,25.508 53.457,25.712 53.553,25.8C53.649,25.888 53.769,25.932 53.913,25.932C54.073,25.932 54.221,25.916 54.357,25.884C54.501,25.852 54.625,25.812 54.729,25.764L54.921,26.808C54.729,26.888 54.501,26.956 54.237,27.012C53.973,27.068 53.737,27.096 53.529,27.096C53.081,27.096 52.729,26.972 52.473,26.724C52.217,26.468 52.089,26.116 52.089,25.668V18.24ZM55.699,18.24H57.019V25.188C57.019,25.508 57.067,25.712 57.162,25.8C57.258,25.888 57.379,25.932 57.522,25.932C57.682,25.932 57.831,25.916 57.966,25.884C58.111,25.852 58.235,25.812 58.339,25.764L58.53,26.808C58.339,26.888 58.111,26.956 57.847,27.012C57.583,27.068 57.347,27.096 57.139,27.096C56.69,27.096 56.339,26.972 56.083,26.724C55.826,26.468 55.699,26.116 55.699,25.668V18.24Z"
-      android:fillColor="#FB7D8A"/>
-  <path
-      android:strokeWidth="1"
-      android:pathData="M21,11.5L57,11.5A7.5,7.5 0,0 1,64.5 19L64.5,27A7.5,7.5 0,0 1,57 34.5L21,34.5A7.5,7.5 0,0 1,13.5 27L13.5,19A7.5,7.5 0,0 1,21 11.5z"
-      android:strokeAlpha="0.25"
-      android:fillColor="#00000000"
-      android:strokeColor="#FB7D8A"/>
-</vector>
